# web/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACLMR Labour Market API",
    "description": "Typed JSON over the local job-ads aggregates. Reads only derived parquet \u2014 never the upstream corpus at request time.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/overview": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Overview",
        "operationId": "overview_api_overview_get",
        "parameters": [
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/series/postings": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Series Postings",
        "operationId": "series_postings_api_series_postings_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(index|level|yoy)$",
              "default": "index",
              "title": "Metric"
            }
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesPoint"
                  },
                  "title": "Response Series Postings Api Series Postings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rank/{dim}": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Rank",
        "operationId": "rank_api_rank__dim__get",
        "parameters": [
          {
            "name": "dim",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dim"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(value|yoy)$",
              "default": "value",
              "title": "Order"
            }
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RankItem"
                  },
                  "title": "Response Rank Api Rank  Dim  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/geography": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Geography",
        "operationId": "geography_api_geography_get",
        "parameters": [
          {
            "name": "measure",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(per10k|lq|count)$",
              "default": "per10k",
              "title": "Measure"
            }
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/wages": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Wages",
        "operationId": "wages_api_wages_get",
        "parameters": [
          {
            "name": "dim",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(occupation|province|overall)$",
              "default": "occupation",
              "title": "Dim"
            }
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WagesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skills": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Skills",
        "operationId": "skills_api_skills_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(top|distinctive)$",
              "default": "top",
              "title": "Mode"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 15,
              "title": "Limit"
            }
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements": {
      "get": {
        "tags": [
          "read"
        ],
        "summary": "Requirements",
        "operationId": "requirements_api_requirements_get",
        "parameters": [
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province code (e.g. ON) or 'All Canada'.",
              "title": "Geo"
            },
            "description": "Province code (e.g. ON) or 'All Canada'."
          },
          {
            "name": "occ",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NOC broad label or 'All occupations'.",
              "title": "Occ"
            },
            "description": "NOC broad label or 'All occupations'."
          },
          {
            "name": "ind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NAICS broad label or 'All industries'.",
              "title": "Ind"
            },
            "description": "NAICS broad label or 'All industries'."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start 'YYYY-MM'.",
              "title": "Start"
            },
            "description": "Window start 'YYYY-MM'."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end / as-of 'YYYY-MM'.",
              "title": "End"
            },
            "description": "Window end / as-of 'YYYY-MM'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeographyResponse": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "measure": {
            "type": "string",
            "title": "Measure"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GeoItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "as_of",
          "measure",
          "items"
        ],
        "title": "GeographyResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OverviewResponse": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "kpis": {
            "$ref": "#/components/schemas/Kpis"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "type": "array",
            "title": "Series"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points"
          },
          "top_growing": {
            "items": {
              "$ref": "#/components/schemas/RankItem"
            },
            "type": "array",
            "title": "Top Growing"
          },
          "top_cooling": {
            "items": {
              "$ref": "#/components/schemas/RankItem"
            },
            "type": "array",
            "title": "Top Cooling"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "as_of",
          "kpis",
          "series",
          "key_points",
          "top_growing",
          "top_cooling"
        ],
        "title": "OverviewResponse"
      },
      "RankItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "yoy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yoy"
          },
          "share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "value"
        ],
        "title": "RankItem"
      },
      "RequirementsResponse": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "remote": {
            "items": {
              "$ref": "#/components/schemas/CategoryShare"
            },
            "type": "array",
            "title": "Remote"
          },
          "employment_type": {
            "items": {
              "$ref": "#/components/schemas/CategoryShare"
            },
            "type": "array",
            "title": "Employment Type"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/CategoryShare"
            },
            "type": "array",
            "title": "Education"
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/CategoryShare"
            },
            "type": "array",
            "title": "Experience"
          },
          "language": {
            "items": {
              "$ref": "#/components/schemas/CategoryShare"
            },
            "type": "array",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "as_of",
          "n",
          "remote",
          "employment_type",
          "education",
          "experience",
          "language"
        ],
        "title": "RequirementsResponse"
      },
...
      "SeriesPoint": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "postings": {
            "type": "integer",
            "title": "Postings"
          },
          "index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index"
          },
          "yoy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yoy"
          },
          "series": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "month",
          "postings"
        ],
        "title": "SeriesPoint"
      },
...
      "SkillsResponse": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "as_of",
          "mode",
          "n",
          "items"
        ],
        "title": "SkillsResponse"
      },
...
      "WagesResponse": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "dim": {
            "type": "string",
            "title": "Dim"
          },
          "min_sample": {
            "type": "integer",
            "title": "Min Sample"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WageItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "as_of",
          "dim",
          "min_sample",
          "items"
        ],
        "title": "WagesResponse"
      }
    }
  }
}